week_2_data_ingestion/airflow/requirements.txt
apache-airflow-providers-google
pyarrow
gdown
jinja2